Error handling. Fixed add recipe. Lots of minor changes

+++ kitchenpointers/src/main/java/com/kitchenpointers/controller/RecipeController.java
@@ -1,18 +1,26 @@
 package com.kitchenpointers.controller;
 
+import java.io.IOException;
+
+import javax.servlet.http.HttpServletResponse;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.kitchenpointers.domain.GetRecipesResponse;
 import com.kitchenpointers.domain.Recipe;
 import com.kitchenpointers.domain.SearchCriteria;
 import com.kitchenpointers.service.RecipeService;
+
+import exception.AddRecipeException;
+import exception.NoRecipesFoundException;
 
 @RestController
 @EnableAutoConfiguration
@@ -21,27 +29,30 @@ public class RecipeController {
     @Autowired
     private RecipeService recipeService;
 
-    @RequestMapping(value = "/hello", method = RequestMethod.GET)
-    @ResponseBody
-    public String hello() {
-        System.out.println("SUP");
-        return "hellosasd";
-    }
-
     @RequestMapping(value = "/getRecipeById/{recipeId}", method = RequestMethod.GET)
-    public Recipe getRecipeById(@PathVariable int recipeId) {
+    public Recipe getRecipeById(@PathVariable Integer recipeId) throws NoRecipesFoundException {
         System.out.println("Called get Recipe By ID: " + recipeId);
-
-        return recipeService.getRecipeById(recipeId);
+        
+        Recipe recipe = recipeService.getRecipeById(recipeId);
+        
+        if (recipe == null) {
+        	throw new NoRecipesFoundException("The requested recipeId does not exist");
+        }
+        else {
+        	return recipe;
+        }
     }
     
     @RequestMapping(value = "/getRecipes", method = RequestMethod.POST)
     public GetRecipesResponse getRecipes(@RequestBody SearchCriteria criteria) {
         System.out.println("Called get Recipes!");
-        System.out.println(criteria.getCalories());
 
         GetRecipesResponse recipes = new GetRecipesResponse();
         recipes.setRecipes(recipeService.getRecipes(criteria));
+        
+        if (recipes.getRecipes() == null || recipes.getRecipes().isEmpty()) {
+        	throw new NoRecipesFoundException("No recipes found matching your criteria");
+        }
 
         return recipes;
     }
@@ -49,20 +60,57 @@ public class RecipeController {
     @RequestMapping(value = "/addRecipe", method = RequestMethod.POST)
     public Recipe addRecipe(@RequestBody Recipe recipe) {
         System.out.println("Called add recipe");
+        
+        String checkArgumentResponse = recipeService.checkAddRecipeArguments(recipe);
+        
+        if (checkArgumentResponse != null) {
+        	//Throw a Bad Request if not all required fields are there
+        	throw new IllegalArgumentException(checkArgumentResponse);
+        }
 
-        recipeService.addRecipe(recipe);
-
-        return recipe;
+        int recipeId = recipeService.addRecipe(recipe);
+        
+        Recipe createdRecipe = recipeService.getRecipeById(recipeId);
+        
+        //See if the recipe exists after creation
+    	if (createdRecipe == null) {
+        	throw new AddRecipeException("Error finding recipe after creation");
+        }
+        
+        return createdRecipe;
     }
 
-    @RequestMapping(value = "/deleteRecipe", method = RequestMethod.POST)
-    public int addRecipe(@RequestBody Integer recipeId) {
+    @RequestMapping(value = "/deleteRecipe", method = RequestMethod.DELETE)
+    public String addRecipe(@RequestBody Integer recipeId) {
 
+    	if (recipeId == null) {
+    		throw new IllegalArgumentException("recipeId is required");
+    	}
+    	
     	System.out.println("Called delete with recipeId: " + recipeId);
+    	
+    	//See if the recipe exists
+    	if (recipeService.getRecipeById(recipeId) == null) {
+        	throw new NoRecipesFoundException("Recipe with id '" + recipeId + "' does not exist");
+        }
     	
         recipeService.deleteRecipe(recipeId);
 
-        return recipeId;
+        return "Deleted recipe with ID " + recipeId;
     }
-
+    
+    @ExceptionHandler
+    void handleIllegalArgumentException(IllegalArgumentException e, HttpServletResponse response) throws IOException {
+        response.sendError(HttpStatus.BAD_REQUEST.value());
+    }
+    
+	@ExceptionHandler
+	void handleNoRecipesFoundException(NoRecipesFoundException e, HttpServletResponse response) throws IOException {
+	    response.sendError(HttpStatus.NOT_FOUND.value());
+	}
+	
+	@ExceptionHandler
+	void handleAddRecipeException(AddRecipeException e, HttpServletResponse response) throws IOException {
+	    response.sendError(HttpStatus.EXPECTATION_FAILED.value());
+	}
 }
